Server/requirements.txt
fastapi
uvicorn
firebase_admin
requests
email-validator
motor[srv]
python-dotenv